Reduce the executable size
In release mode, the executable size should be kept to the bare minimum
in order to be used in environments with low storage and reduce the data
consumption when pre-built packages are downloaded. Because performance
is absolutely not needed, it allows to a quite aggressive compiling
strategy (opt-level = 'z').

https://doc.rust-lang.org/cargo/reference/manifest.html#the-profile-sections
Fixes #1