Refactor the account management
Until now, the account management was archaic and it was impossible to
improve it without this heavy refactoring. Accounts are now disjoint
from both certificates and endpoints. They have their own hooks and
their own environment variables. They are stored in a binary file
instead of the PEM exports of the private and public keys.

This refactoring will allow account management to evolve into something
more serious, with real key rollover, contact information update and so
on.